Create database dynamically
Add a class DynamicMarkovChain that inherits from MarkovChain. The main
difference here is that, while an instance of MarkovChain creates a
complete database of words and probabilities and loads it into memory,
DynamicMarkovChain instances only create parts of the database as they
are required. This generally makes the sentence generation slower, but
it can be useful in situations where there is little memory available.

All changes are under-the-hood. If you are a developer using this
library, you will be able to use the DynamicMarkovChain class in the
same way as the MarkovChain class, that is

    mc = MarkovChain()
    mc.generateDatabase(textSample)
    mc.generateString()
    mc.generateStringWithSeed("python")

should produce an equivalent (though not necessarily the same) result as

    dmc = DynamicMarkovChain()
    dmc.generateDatabase(textSample)
    dmc.generateString()
    dmc.generateStringWithSeed("python")

However, you will probably be able to register changes in execution time
and resource usage.
Additionally, DynamicMarkovChain instances provide the method
flushDatabase(), which resets the entire database to free up memory.

The original MarkovChain class was not modified.